# docmind-frontend/DocMind/swagger_spec.json
{"openapi":"3.1.0","info":{"title":"DocMind - AI Document Intelligence & RAG API","description":"Rest API for DocMind: Multi-Format document ingestion, vector embeddings with Qdrant and hybrid conversational Q&A with Nvidia.","contact":{"name":"Pranit Bhangale","url":"https://pranitbhangale.vercel.app","email":"pranitbhangale1453@gmail.com"},"version":"1.0.0","summary":"DocMind API"},"servers":[{"url":"http://localhost:8080","description":"Generated server url"}],"tags":[{"name":"Document Management","description":"Endpoints for uploading, retrieving, listing, and deleting documents and their vector embeddings."},{"name":"Account","description":"User account management APIs"},{"name":"Registration Management","description":"Endpoints for registering new users and administrators."},{"name":"Password Management","description":"Endpoints for requesting password resets, resetting forgotten passwords, and changing account passwords."},{"name":"Chat Management","description":"Endpoints for querying documents using AI-powered chat, including standard and streaming responses."},{"name":"Authentication","description":"Endpoints for user authentication, token refresh, and logout."},{"name":"Verification Management","description":"Endpoints for verifying OTPs for account registration and password reset."},{"name":"Admin Portal","description":"Endpoints for managing users and their role assignments."}],"paths":{"/api/verify":{"post":{"tags":["Verification Management"],"summary":"Verify registration OTP","description":"Verifies the OTP provided during user registration and activates the user account.","operationId":"verifyRegistration","parameters":[{"name":"X-API-Version","in":"header","schema":{"type":"string","default":"1","enum":["v1"]}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/VerifyOtp"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/VerificationResponse"}}}}}}},"/api/verify/reset":{"post":{"tags":["Verification Management"],"summary":"Verify password reset OTP","description":"Verifies the OTP provided for password reset and authorizes the password reset process.","operationId":"verifyPasswordResetOtp","parameters":[{"name":"X-API-Version","in":"header","schema":{"type":"string","default":"1","enum":["v1"]}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/VerifyOtp"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/VerificationResponse"}}}}}}},"/api/register":{"post":{"tags":["Registration Management"],"summary":"Register new user","description":"Creates a new user account and initiates the account verification process.","operationId":"registerNewUser","parameters":[{"name":"X-API-Version","in":"header","schema":{"type":"string","default":"1","enum":["v1"]}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/SignupRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/SignupResponse"}}}}}}},"/api/register/admin":{"post":{"tags":["Registration Management"],"summary":"Register new administrator","description":"Creates a new administrator account with the required administrative privileges.","operationId":"registerNewAdmin","parameters":[{"name":"X-API-Version","in":"header","schema":{"type":"string","default":"1","enum":["v1"]}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/SignupRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/SignupResponse"}}}}}}},"/api/password/reset/request":{"post":{"tags":["Password Management"],"summary":"Request password reset","description":"Initiates the password reset process by sending a password reset OTP to the user's registered email address.","operationId":"requestPasswordReset","parameters":[{"name":"X-API-Version","in":"header","schema":{"type":"string","default":"1","enum":["v1"]}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ForgotPasswordEmail"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/PasswordResponse"}}}}}}},"/api/documents/upload":{"post":{"tags":["Document Management"],"summary":"Upload a document","description":"Uploads a document and processes it for vector embedding.","operationId":"uploadDocument","parameters":[{"name":"X-API-Version","in":"header","schema":{"type":"string","default":"1","enum":["v1"]}}],"requestBody":{"content":{"multipart/form-data":{"schema":{"type":"object","properties":{"file":{"type":"string","format":"binary"}},"required":["file"]}}}},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiResponseDocumentResponse"}}}}}}},"/api/chat/documents/{documentId}/query":{"post":{"tags":["Chat Management"],"summary":"Ask a question about a document","description":"Sends a question to the AI assistant and returns a response based on the specified document and conversation context.","operationId":"getResponse","parameters":[{"name":"documentId","in":"path","required":true,"schema":{"type":"string","format":"uuid"}},{"name":"X-Conversation-ID","in":"header","required":true,"schema":{"type":"string","format":"uuid"}},{"name":"X-API-Version","in":"header","schema":{"type":"string","default":"1","enum":["v1"]}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/QueryRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/QueryResponse"}}}}}}},"/api/chat/documents/{documentId}/query/stream":{"post":{"tags":["Chat Management"],"summary":"Stream a response about a document","description":"Sends a question to the AI assistant and streams the generated response incrementally using Server-Sent Events (SSE).","operationId":"streamChat","parameters":[{"name":"documentId","in":"path","required":true,"schema":{"type":"string","format":"uuid"}},{"name":"X-Conversation-ID","in":"header","required":true,"schema":{"type":"string","format":"uuid"}},{"name":"X-API-Version","in":"header","schema":{"type":"string","default":"1","enum":["v1"]}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/QueryRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"text/event-stream":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/ServerSentEventString"}}}}}}}},"/api/auth/refresh":{"post":{"tags":["Authentication"],"summary":"Refresh access token","description":"Validates the refresh token and generates a new access and refresh token.","operationId":"refreshToken","parameters":[{"name":"X-API-Version","in":"header","schema":{"type":"string","default":"1","enum":["v1"]}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/TokenResponse"}}}}}}},"/api/auth/logout":{"post":{"tags":["Authentication"],"summary":"Logout user","description":"Revokes the user's access and refresh tokens and clears the authentication cookies.","operationId":"logout","parameters":[{"name":"X-API-Version","in":"header","schema":{"type":"string","default":"1","enum":["v1"]}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"string"}}}}}}},"/api/auth/login":{"post":{"tags":["Authentication"],"summary":"Authenticate user","description":"Authenticates the user and sets the access and refresh tokens in secure HTTP-only cookies.","operationId":"authenticateUser","parameters":[{"name":"X-API-Version","in":"header","schema":{"type":"string","default":"1","enum":["v1"]}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/LoginRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/LoginRespone"}}}}}}},"/api/admin/portal/user/{userId}/control":{"post":{"tags":["Admin Portal"],"summary":"Control user account","description":"Enables, disables, or updates the account status of a user.","operationId":"userAccountControl","parameters":[{"name":"userId","in":"path","required":true,"schema":{"type":"string","format":"uuid"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserAccountControlRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiResponseVoid"}}}}}}},"/api/admin/portal/assign":{"post":{"tags":["Admin Portal"],"summary":"Assign role to user","description":"Assigns a role to the specified user.","operationId":"assignRoleToUser","parameters":[{"name":"X-API-Version","in":"header","schema":{"type":"string","default":"1","enum":["v1"]}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/AssignUserRoleRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/AssignResponse"}}}}}}},"/api/account/deactivate":{"post":{"tags":["Account"],"summary":"Deactivate user account","description":"Deactivates the authenticated user's account and schedules permanent deletion after 15 days.","operationId":"deactivateUserAccount","parameters":[{"name":"X-API-Version","in":"header","schema":{"type":"string","default":"1","enum":["v1"]}}],"responses":{"200":{"description":"OK"}}}},"/api/password/reset":{"patch":{"tags":["Password Management"],"summary":"Reset forgotten password","description":"Resets the user's password after successful verification of the password reset request.","operationId":"resetPassword","parameters":[{"name":"X-API-Version","in":"header","schema":{"type":"string","default":"1","enum":["v1"]}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ChangeForgotPasswordRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/PasswordResponse"}}}}}}},"/api/password/change":{"patch":{"tags":["Password Management"],"summary":"Change account password","description":"Changes the authenticated user's current account password.","operationId":"changeAccountPassword","parameters":[{"name":"X-API-Version","in":"header","schema":{"type":"string","default":"1","enum":["v1"]}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ChangePassword"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ChangePasswordResponse"}}}}}}},"/api/admin/portal/{userRoleId}/revoke":{"patch":{"tags":["Admin Portal"],"summary":"Revoke role assigned to user","description":"Revokes an active role assignment from a user.","operationId":"revokeRoleAssignedToUser","parameters":[{"name":"userRoleId","in":"path","required":true,"schema":{"type":"integer","format":"int64"}},{"name":"X-API-Version","in":"header","schema":{"type":"string","default":"1","enum":["v1"]}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/RevokeUserRoleRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/RevokeResponse"}}}}}}},"/api/admin/portal/{userRoleId}/reassign":{"patch":{"tags":["Admin Portal"],"summary":"Re-assign role to user","description":"Re-activates a previously inactive role assignment.","operationId":"reAssignRoleToUser","parameters":[{"name":"userRoleId","in":"path","required":true,"schema":{"type":"integer","format":"int64"}},{"name":"X-API-Version","in":"header","schema":{"type":"string","default":"1","enum":["v1"]}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/AssignUserRoleRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/AssignResponse"}}}}}}},"/api/documents":{"get":{"tags":["Document Management"],"summary":"List documents","description":"Retrieves a paginated list of documents with optional keyword search and sorting.","operationId":"fetchAllDocuments","parameters":[{"name":"keyword","in":"query","required":false,"schema":{"type":"string"}},{"name":"scrollId","in":"query","required":false,"schema":{"type":"string"}},{"name":"pageSize","in":"query","required":false,"schema":{"type":"integer","format":"int32","default":5,"maximum":5,"minimum":1}},{"name":"sortBy","in":"query","required":false,"schema":{"type":"string","default":"fileName"}},{"name":"sortDirection","in":"query","required":false,"schema":{"type":"string","default":"ASC"}},{"name":"scrollDirection","in":"query","required":false,"schema":{"type":"string","default":"FORWARD"}},{"name":"X-API-Version","in":"header","schema":{"type":"string","default":"1","enum":["v1"]}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ScrollResponseDocumentResponse"}}}}}}},"/api/documents/{documentId}":{"get":{"tags":["Document Management"],"summary":"Get document by ID","description":"Retrieves a document and its metadata using its unique ID.","operationId":"fetchDocumentById","parameters":[{"name":"documentId","in":"path","required":true,"schema":{"type":"string","format":"uuid"}},{"name":"X-API-Version","in":"header","schema":{"type":"string","default":"1","enum":["v1"]}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiResponseDocumentResponse"}}}}}}},"/api/admin/portal/{userRoleId}":{"get":{"tags":["Admin Portal"],"summary":"Get user role assignment by ID","description":"Returns the user and role details for the specified user-role assignment.","operationId":"fetchUserRoleById","parameters":[{"name":"userRoleId","in":"path","required":true,"schema":{"type":"integer","format":"int64"}},{"name":"X-API-Version","in":"header","schema":{"type":"string","default":"1","enum":["v1"]}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/UserRoleResponse"}}}}}}},"/api/admin/portal/all/users":{"get":{"tags":["Admin Portal"],"summary":"Get all users","description":"Returns a paginated list of users with their assigned roles.","operationId":"fetchAllUsers","parameters":[{"name":"page","in":"query","required":false,"schema":{"type":"integer","format":"int32","default":0,"minimum":0}},{"name":"size","in":"query","required":false,"schema":{"type":"integer","format":"int32","default":10,"maximum":10,"minimum":1}},{"name":"keyword","in":"query","required":false,"schema":{"type":"string"}},{"name":"sortBy","in":"query","required":false,"schema":{"type":"string","default":"username"}},{"name":"sortDirection","in":"query","required":false,"schema":{"type":"string","default":"ASC"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/PageResponseUserResponse"}}}}}}},"/api/documents/{documentId}/delete":{"delete":{"tags":["Document Management"],"summary":"Delete a document","description":"Deletes a document and its associated vector embeddings using its unique ID.","operationId":"deleteDocumentById","parameters":[{"name":"documentId","in":"path","required":true,"schema":{"type":"string","format":"uuid"}},{"name":"X-API-Version","in":"header","schema":{"type":"string","default":"1","enum":["v1"]}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiResponseVoid"}}}}}}}},"components":{"schemas":{"VerifyOtp":{"type":"object","properties":{"challengeId":{"type":"string","minLength":1},"otp":{"type":"string","minLength":1}},"required":["challengeId","otp"]},"VerificationResponse":{"type":"object","properties":{"message":{"type":"string"}}},"SignupRequest":{"type":"object","properties":{"fullName":{"type":"string","minLength":1},"username":{"type":"string","maxLength":50,"minLength":3},"email":{"type":"string","format":"email","minLength":1},"password":{"type":"string","maxLength":100,"minLength":8,"pattern":"^(?=.*[A-Z])(?=.*[a-z])(?=.*\\d)(?=.*[@#$%^&+=!]).{8,}$"},"confirmPassword":{"type":"string","minLength":1}},"required":["confirmPassword","email","fullName","password","username"]},"SignupResponse":{"type":"object","properties":{"challengeId":{"type":"string"},"message":{"type":"string"}}},"ForgotPasswordEmail":{"type":"object","properties":{"email":{"type":"string","format":"email","minLength":1}},"required":["email"]},"PasswordResponse":{"type":"object","properties":{"challengeId":{"type":"string"},"message":{"type":"string"}}},"ApiResponseDocumentResponse":{"type":"object","properties":{"status":{"type":"boolean"},"message":{"type":"string"},"data":{"$ref":"#/components/schemas/DocumentResponse"},"timestamp":{"type":"string","format":"date-time"}}},"DocumentResponse":{"type":"object","properties":{"documentId":{"type":"string","format":"uuid"},"fileName":{"type":"string"},"fileSize":{"type":"integer","format":"int64"},"status":{"type":"string","enum":["UPLOADING","PROCESSING","INDEXED","FAILED"]},"chunksCreated":{"type":"integer","format":"int64"},"createdAt":{"type":"string","format":"date-time"}}},"QueryRequest":{"type":"object","properties":{"query":{"type":"string","minLength":1},"provider":{"type":"string","enum":["NVIDIA"]},"options":{"$ref":"#/components/schemas/RetrievalOptions"}},"required":["provider","query"]},"RetrievalOptions":{"type":"object","properties":{"topK":{"type":"integer","format":"int32","maximum":10,"minimum":1},"similarityThreshold":{"type":"number","format":"double","maximum":1.0,"minimum":0.0}}},"QueryResponse":{"type":"object","properties":{"content":{"type":"string"}}},"ServerSentEventString":{},"TokenResponse":{"type":"object","properties":{"message":{"type":"string"}}},"LoginRequest":{"type":"object","properties":{"username":{"type":"string","minLength":1},"password":{"type":"string","minLength":1}},"required":["password","username"]},"LoginRespone":{"type":"object","properties":{"username":{"type":"string"},"roles":{"type":"array","items":{"type":"string"},"uniqueItems":true}}},"UserAccountControlRequest":{"type":"object","properties":{"enabled":{"type":"boolean"}}},"ApiResponseVoid":{"type":"object","properties":{"status":{"type":"boolean"},"message":{"type":"string"},"data":{},"timestamp":{"type":"string","format":"date-time"}}},"AssignUserRoleRequest":{"type":"object","properties":{"userId":{"type":"string","format":"uuid"},"roleId":{"type":"integer","format":"int64"}},"required":["roleId","userId"]},"AssignResponse":{"type":"object","properties":{"status":{"type":"boolean"},"message":{"type":"string"}}},"ChangeForgotPasswordRequest":{"type":"object","properties":{"email":{"type":"string","format":"email","minLength":1},"passwordRequest":{"$ref":"#/components/schemas/PasswordRequest"}},"required":["email"]},"PasswordRequest":{"type":"object","properties":{"password":{"type":"string","minLength":1},"confirmPassword":{"type":"string","minLength":1}},"required":["confirmPassword","password"]},"ChangePassword":{"type":"object","properties":{"oldPassword":{"type":"string","minLength":1},"request":{"$ref":"#/components/schemas/PasswordRequest"}},"required":["oldPassword"]},"ChangePasswordResponse":{"type":"object","properties":{"message":{"type":"string"}}},"RevokeUserRoleRequest":{"type":"object","properties":{"userId":{"type":"string","format":"uuid"},"roleId":{"type":"integer","format":"int64","minimum":0}},"required":["roleId","userId"]},"RevokeResponse":{"type":"object","properties":{"status":{"type":"boolean"},"message":{"type":"string"}}},"ScrollResponseDocumentResponse":{"type":"object","properties":{"hasPrevious":{"type":"boolean"},"prevScrollId":{"type":"string"},"contents":{"type":"array","items":{"$ref":"#/components/schemas/DocumentResponse"}},"hasNext":{"type":"boolean"},"nextScrollId":{"type":"string"},"pageSize":{"type":"integer","format":"int32"}}},"RoleResponses":{"type":"object","properties":{"roleId":{"type":"integer","format":"int64"},"roleName":{"type":"string"},"roleDescription":{"type":"string"},"status":{"type":"string","enum":["ACTIVE","INACTIVE"]}}},"UserRoleResponse":{"type":"object","properties":{"userRoleId":{"type":"integer","format":"int64"},"userId":{"type":"string","format":"uuid"},"username":{"type":"string"},"roles":{"type":"array","items":{"$ref":"#/components/schemas/RoleResponses"}}}},"PageResponseUserResponse":{"type":"object","properties":{"contents":{"type":"array","items":{"$ref":"#/components/schemas/UserResponse"}},"currentPage":{"type":"integer","format":"int32"},"pageSize":{"type":"integer","format":"int32"},"totalElements":{"type":"integer","format":"int64"},"totalPages":{"type":"integer","format":"int32"},"isLastPage":{"type":"boolean"},"isFirstPage":{"type":"boolean"}}},"RoleResponse":{"type":"object","properties":{"roleId":{"type":"integer","format":"int64"},"roleName":{"type":"string"},"roleDescription":{"type":"string"},"status":{"type":"string","enum":["ACTIVE","INACTIVE"]}}},"UserResponse":{"type":"object","properties":{"userId":{"type":"string","format":"uuid"},"username":{"type":"string"},"email":{"type":"string"},"enabled":{"type":"boolean"},"deleted":{"type":"boolean"},"roles":{"type":"array","items":{"$ref":"#/components/schemas/RoleResponse"}}}}},"securitySchemes":{"accessToken":{"type":"apiKey","description":"JWT access token stored in HttpOnly cookie","name":"access_token","in":"cookie"}}}}